Get rid of backend structure.
This is a huge diff that eliminates the use of the Backend structure.
There were significant layer violations that require the code to be
flattened before it can be fixed.

The next few steps include making user, smtp, piwww etc either API calls
or packages.

While this diff is not remotely complete we are moving it into master so
that more people can work on it at the same time.